# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ONUSLY API
  description: Accountability and single-approver goal-verification REST API for Phase 1.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Returns service availability status.
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/google:
    post:
      summary: Authenticate with Google
      description: Exchanges a Google ID Token / credential for an ONUSLY JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idToken:
                  type: string
                credential:
                  type: string
      responses:
        '200':
          description: Successful authentication.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOi...
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body.
        '401':
          description: Invalid Google token.

  /goals:
    post:
      summary: Create a new goal
      security:
        - BearerAuth: []
      description: Creates an active goal belonging to the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: Complete Chapter 4
                description:
                  type: string
                  example: Read and solve exercises
                blockedApps:
                  type: array
                  items:
                    type: string
                  example: ["com.instagram.android", "com.twitter.android"]
                blockedDomains:
                  type: array
                  items:
                    type: string
                  example: ["instagram.com", "x.com"]
      responses:
        '201':
          description: Goal created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          description: Title is required, invalid package name or domain, or exceeding 20 items.
        '401':
          description: Missing or invalid JWT.

    get:
      summary: List user goals
      security:
        - BearerAuth: []
      description: Returns all goals owned by the authenticated caller.
      responses:
        '200':
          description: List of goals.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Goal'
        '401':
          description: Missing or invalid JWT.

  /goals/{id}:
    get:
      summary: Get goal by ID
      security:
        - BearerAuth: []
      description: Returns a specific goal owned by the authenticated caller.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Goal retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '401':
          description: Missing or invalid JWT.
        '404':
          description: Goal not found or belongs to another user.

    patch:
      summary: Update goal details, accountability, or targets
      security:
        - BearerAuth: []
      description: |
        Updates goal title, description, accountability settings, or targets.
        Targets are add-only: removing existing targets is prohibited.
        Changing accountability to none while targets are locked is prohibited.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                approvalType:
                  type: string
                  enum: [community, friend, none]
                approverEmail:
                  type: string
                blockedApps:
                  type: array
                  items:
                    type: string
                blockedDomains:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Goal updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          description: Validation error or goal not active.
        '401':
          description: Missing or invalid JWT.
        '404':
          description: Goal not found.
        '409':
          description: Target locking conflict (targets_locked, goal_locked, or changing accountability to none rejected).

    delete:
      summary: Delete goal
      security:
        - BearerAuth: []
      description: Deletes a goal owned by the authenticated caller. Goals with targets cannot be deleted unless completed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Goal deleted.
        '401':
          description: Missing or invalid JWT.
        '409':
          description: Goal locked (goal_locked) - active goal with targets cannot be deleted.

  /goals/{id}/proofs:
    post:
      summary: Submit proof for an active goal
      security:
        - BearerAuth: []
      description: |
        Submits proof and atomically transitions the goal from active to proof_submitted.
        DynamoDB conditional transactions prevent duplicate submissions or submissions against non-active goals.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - proofType
              properties:
                proofType:
                  type: string
                  enum: [text, link, photo]
                textExplanation:
                  type: string
                externalLink:
                  type: string
                photoUrl:
                  type: string
      responses:
        '201':
          description: Proof submitted and goal transitioned to proof_submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proof'
        '400':
          description: Invalid proof type or missing required evidence fields.
        '401':
          description: Missing or invalid JWT.
        '404':
          description: Goal not found.
        '409':
          description: Proof has already been submitted or goal is not active.

    get:
      summary: List proofs for a goal
      security:
        - BearerAuth: []
      description: Returns all proof records submitted for the specified goal.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of proofs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proof'
        '401':
          description: Missing or invalid JWT.

  /goals/{id}/proofs/{proofId}:
    get:
      summary: Get proof by ID
      security:
        - BearerAuth: []
      description: Returns a specific proof by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: proofId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Proof retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proof'
        '401':
          description: Missing or invalid JWT.
        '404':
          description: Proof not found.

  /goals/{id}/proofs/{proofId}/approval:
    post:
      summary: Decide on submitted proof
      security:
        - BearerAuth: []
      description: |
        Assigned approver submits a decision ("approved" or "rejected").
        Atomically records the approval and transitions the goal:
        - approved -> completed
        - rejected -> active
        Protected by DynamoDB conditional transactions ensuring only goals in proof_submitted can be decided.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: proofId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ownerId
                - status
              properties:
                ownerId:
                  type: string
                  description: User ID of the goal owner.
                status:
                  type: string
                  enum: [approved, rejected]
                comment:
                  type: string
      responses:
        '201':
          description: Approval decision recorded and goal status transitioned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          description: Invalid decision status or mismatched proof.
        '401':
          description: Missing or invalid JWT.
        '403':
          description: Forbidden (authenticated user is not the assigned approver, or owner attempted self-approval).
        '404':
          description: Goal or proof not found.
        '409':
          description: Proof has already been decided or goal is not in proof_submitted state.

  /goals/{id}/approvals:
    get:
      summary: List approvals for a goal
      security:
        - BearerAuth: []
      description: Returns decision records for the goal.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of approval decisions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Approval'
        '401':
          description: Missing or invalid JWT.

  /goals/{id}/approvals/{approvalId}:
    get:
      summary: Get approval decision by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approval decision retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '401':
          description: Missing or invalid JWT.
        '404':
          description: Approval not found.

  /enforcement/blocklist:
    get:
      summary: Get active platform enforcement blocklist
      security:
        - BearerAuth: []
      description: |
        Returns the combined active targets (Android package names and web domains) for the authenticated user.
        Includes only goals where status is 'active' or 'proof_submitted' and at least one target exists.
        Supports HTTP conditional requests using the ETag / If-None-Match header.
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
          description: Previous version ETag string (e.g. v1-3fa85f647c9842a1)
      responses:
        '200':
          description: Blocklist retrieved.
          headers:
            ETag:
              schema:
                type: string
              example: '"v1-4f53cda18c2baa0c"'
            Cache-Control:
              schema:
                type: string
              example: 'private, no-cache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlocklistResponse'
        '304':
          description: Not Modified. The client already has the latest target version. No response body.
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
        '401':
          description: Missing or invalid JWT.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        picture:
          type: string
        createdAt:
          type: string

    Goal:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        approverId:
          type: string
        approverEmail:
          type: string
        approvalType:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, proof_submitted, completed]
        createdAt:
          type: string
        updatedAt:
          type: string
        blockedApps:
          type: array
          items:
            type: string
        blockedDomains:
          type: array
          items:
            type: string

    Proof:
      type: object
      properties:
        id:
          type: string
        goalId:
          type: string
        ownerId:
          type: string
        proofType:
          type: string
          enum: [text, link, photo]
        textExplanation:
          type: string
        externalLink:
          type: string
        photoUrl:
          type: string
        submittedAt:
          type: string

    Approval:
      type: object
      properties:
        id:
          type: string
        goalId:
          type: string
        proofId:
          type: string
        ownerId:
          type: string
        approverId:
          type: string
        status:
          type: string
          enum: [approved, rejected]
        comment:
          type: string
        decidedAt:
          type: string

    TargetItem:
      type: object
      properties:
        goalId:
          type: string
        goalTitle:
          type: string
        status:
          type: string
          enum: [active, proof_submitted]
        apps:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string

    BlocklistResponse:
      type: object
      properties:
        userId:
          type: string
        version:
          type: string
          example: v1-4f53cda18c2baa0c
        generatedAt:
          type: string
        targets:
          type: array
          items:
            $ref: '#/components/schemas/TargetItem'